requirements.txt
streamlit==1.32.0
pandas==2.2.0
psycopg2-binary==2.9.9
plotly==5.19.0
